website: packages: Fix list of supported architecture.
package-transitive-supported-systems looks at the transitive closure of
inputs for the current architecture.  When we generate the website on
x86_64, it selects the reduced bootstrap path, which is only supported
on x86_64 and i686.  The result is that it would show only these two
architectures as supported when it is not the case.

Thanks to baryluk on IRC for noticing the set of supported architecture
on the website was so small.

* website/apps/packages/templates/components.scm (package-t): Fix list of
supported architecture.