website: Fix dropdown menu shrinking when clicked.
When clicking or tapping on a dropdown button in the primary menu of
the website the menu width is shrunk to its min-width, and long items
are broken into separate lines making them look as if they were
different items in the menu.

This change fixes these problems by leaving the menu width as wide as
its longest item when the button is toggled.

* apps/base/templates/components.scm (menu-dropdown): Set width to
  max-content on :checked.